requirements.txt
# WebSocket Signal Server Requirements

# Core dependencies
asyncpg>=0.27.0      # PostgreSQL async driver
websockets>=11.0     # WebSocket server/client
python-dotenv>=1.0.0 # Environment variables

# Optional dependencies for production
uvloop>=0.17.0       # Faster event loop
aiofiles>=23.0.0     # Async file operations for logging

# Testing and monitoring
pytest>=7.4.0
pytest-asyncio>=0.21.0
aiohttp>=3.8.0       # For health check endpoint